refactor: 리뷰 반영
- 브랜드 도메인과 DTO에 원본 이미지 경로 필드 생성
- interface를 사용한 스토리지 서비스 추상화

issue number: #48
