+++ src/main/java/com/flab/shoeauction/service/storage/StorageService.java
@@ -1,0 +1,8 @@
+package com.flab.shoeauction.service.storage;
+
+import org.springframework.web.multipart.MultipartFile;
+
+public interface StorageService {
+
+    String upload(MultipartFile file, String destLocation);
+}
